Change time and time_bnds for climate forcing
Now we use the middle of a month for the the time
and the first and last day as bounds. This is now using a 360 day
calendar to make this easy. Before that, I used 365 day calendar and the
time was always the first day of the month, but this did not always work
because months had different number of days